ScyllaDB/pyproject.toml
[project]
name = "scylladb-dev"
version = "0.1.0"
description = "ScyllaDB local development environment with Python support"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "cassandra-driver>=3.29.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
